refactor: split client handlers into separate files
